backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=4.2.0",
    "fastapi>=0.128.0",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "stytch>=14.1.0",
    "svix>=1.84.1",
    "uvicorn>=0.40.0",
]
